api/requirements.txt
accelerate==0.30.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiopg==1.4.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Authlib==1.3.1
backcall==0.2.0
beautifulsoup4==4.12.0
black==23.12.1
bleach==6.0.0
blinker==1.8.2
build==0.10.0
CacheControl==0.12.11
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.0.1
click==8.1.7
cloud-sql-python-connector==1.12.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
crashtest==0.4.1
cryptography==43.0.1
cssbeautifier==1.14.8
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
diff_cover==8.0.3
distlib==0.3.6
dnspython==2.6.1
docstring_parser==0.16
dulwich==0.21.5
EditorConfig==0.12.3
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.113.0
fastapi-cli==0.0.5
fastjsonschema==2.16.3
filelock==3.12.0
Flask==3.0.3
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.19.0
google-auth==2.34.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==1.42.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gotrue==2.8.1
greenlet==3.1.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.64.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.23.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.6.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.2.3
javascript==1!1.0.1
jedi==0.18.2
Jinja2==3.1.4
joblib==1.4.2
jsbeautifier==1.14.8
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
keyring==23.13.1
kiwisolver==1.4.4
llvmlite==0.42.0
lockfile==0.12.2
mailersend==0.5.8
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.3
next==1.0.1
node==1.2.1
nodejs==0.1.1
notebook==6.5.3
notebook_shim==0.2.2
npm==0.1.1
npx==0.1.1
numba==0.59.1
numpy==1.26.4
odict==1.9.0
opencv-contrib-python==4.7.0.72
openpyxl==3.1.2
optional-django==0.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pg8000==1.31.2
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.1
plotly==5.22.0
pluggy==1.5.0
plumber==1.7
poetry==1.5.0
poetry-core==1.6.0
poetry-plugin-export==1.3.1
postgrest==0.16.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==5.27.1
psutil==5.9.4
psycopg==3.1.10
psycopg-pool==3.1.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.9.0
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
pyflakes==3.0.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pynndescent==0.5.12
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.0.2
q==2.7
qtconsole==5.4.1
QtPy==2.3.0
rapidfuzz==2.15.1
realtime==2.0.2
regex==2023.6.3
requests==2.32.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.2
scramp==1.4.5
Send2Trash==1.8.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.4
SQLAlchemy==2.0.35
sqlfluff==2.3.5
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.38.4
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.12
tblib==3.0.0
tenacity==8.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
torc==0.1.1
torch==2.3.0
torchvision==0.18.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trove-classifiers==2023.5.2
twilio==9.0.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
uri-template==1.2.0
urllib3==2.2.1
uvicorn==0.30.6
uvloop==0.20.0
validators==0.28.3
virtualenv==20.23.0
watchfiles==0.24.0
wcwidth==0.2.6
weaviate-client==4.6.4
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.6
workos==2.1.0
xattr==0.10.1
yarl==1.9.4
zipp==3.15.0
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
